Rename Activities and Fragments.
