+++ app/src/main/java/jp/nk5/saifu2/AccountActivity.java
@@ -3,11 +3,11 @@ package jp.nk5.saifu2;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class BankActivity extends AppCompatActivity {
+public class AccountActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_bank);
+        setContentView(R.layout.activity_account);
     }
 }
