+++ app/src/main/java/jp/nk5/saifu2/TopActivity.java
@@ -5,24 +5,23 @@ import android.os.Bundle;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.AppCompatActivity;
-import android.widget.Toast;
 
-import jp.nk5.saifu2.fragment.MainFragment;
-import jp.nk5.saifu2.fragment.MenuFragment;
+import jp.nk5.saifu2.view.TopFragment;
+import jp.nk5.saifu2.view.TopMenuFragment;
 
-public class MainActivity extends AppCompatActivity implements MenuFragment.EventListener {
+public class TopActivity extends AppCompatActivity implements TopMenuFragment.EventListener {
 
     private FragmentManager fragmentManager;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_top);
 
         fragmentManager = getSupportFragmentManager();
         fragmentManager.beginTransaction()
-                .replace(R.id.layout_menu, new MenuFragment(), MenuFragment.getTagName())
-                .replace(R.id.layout_main, new MainFragment(), MainFragment.getTagName())
+                .replace(R.id.layout_menu, new TopMenuFragment(), TopMenuFragment.getTagName())
+                .replace(R.id.layout_main, new TopFragment(), TopFragment.getTagName())
                 .commit();
     }
 
@@ -30,7 +29,7 @@ public class MainActivity extends AppCompatActivity implements MenuFragment.Even
     {
         DrawerLayout drawer = findViewById(R.id.drawer_layout);
         drawer.closeDrawers();
-        Intent intent = new Intent(this, BankActivity.class);
+        Intent intent = new Intent(this, AccountActivity.class);
         startActivity(intent);
     }
