+++ app/src/main/res/layout/activity_account.xml
@@ -5,7 +5,7 @@
     android:id="@+id/drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="jp.nk5.saifu2.BankActivity">
+    tools:context="jp.nk5.saifu2.AccountActivity">
 
     <FrameLayout
         android:id="@+id/layout_main"
